General source code improvements: Declared a lot of functions and variables as static that are only used in one file. Functions without parameters now use correct ANSI prototypes. Added accelerated memory access functions for systems that can do non-aligned memory access. (Thanks to Eero Tamminen for the huge patch!)

+++ src/includes/file.h
@@ -38,7 +38,7 @@ extern BOOL File_FileNameIsZIP(char *pszFileName);
 extern BOOL File_FileNameIsMSA(char *pszFileName);
 extern BOOL File_FileNameIsST(char *pszFileName);
 extern void *File_Read(char *pszFileName, void *pAddress, long *pFileSize, char *ppszExts[]);
-extern BOOL File_Save(char *pszFileName, void *pAddress,long Size,BOOL bQueryOverwrite);
+extern BOOL File_Save(char *pszFileName, void *pAddress, size_t Size, BOOL bQueryOverwrite);
 extern int File_Length(char *pszFileName);
 extern BOOL File_Exists(char *pszFileName);
 extern BOOL File_Delete(char *pszFileName);
